+++ packages/assets/projects/fallbacks/tests/unit/config/vitest.config.ts
@@ -1,9 +1,14 @@
+import { resolve } from 'node:path';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['**/src/**/*.spec.ts'],
-    globals: true,
+    root: resolve(__dirname, '../../../'),
+    include: ['**/*.spec.ts'],
     environment: 'jsdom',
+    globals: true,
+    typecheck: {
+      tsconfig: './tsconfig.test.json',
+    },
   },
 });