+++ packages/components-js/tests/vrt/specs/common/components.vrt.ts
@@ -3,7 +3,7 @@ import { getComponentMeta } from '@porsche-design-system/component-meta';
 import { TAG_NAMES, type TagName } from '@porsche-design-system/shared';
 import { schemes, themes, viewportWidthM, viewportWidths } from '@porsche-design-system/shared/testing';
 import * as globby from 'globby-legacy';
-import path from 'path';
+import * as path from 'path';
 import { setupScenario } from '../../helpers';
 
 const sourceDirectory = path.resolve('src/pages');
@@ -66,7 +66,7 @@ const revertAutoFocus = async (page: Page, component: string): Promise<void> => 
 };
 
 test('should have certain amount of components', () => {
-  expect(components.length).toBe(70);
+  expect(components.length).toBe(73);
 });
 
 for (const component of components) {