+++ packages/components-react/src/pages/index.tsx
@@ -71,6 +71,7 @@ import { FlagPage } from './generated/Flag';
 import { FlexPage } from './generated/Flex';
 import { Flyout_1Page } from './generated/Flyout_1';
 import { Flyout_2Page } from './generated/Flyout_2';
+import { Flyout_3Page } from './generated/Flyout_3';
 import { FlyoutBackdropBlurPage } from './generated/FlyoutBackdropBlur';
 import { FlyoutBackdropShadingPage } from './generated/FlyoutBackdropShading';
 import { FlyoutBasicPage } from './generated/FlyoutBasic';
@@ -112,6 +113,7 @@ import { IconPage } from './generated/Icon';
 import { InlineNotificationPage } from './generated/InlineNotification';
 import { InputDatePage } from './generated/InputDate';
 import { InputEmailPage } from './generated/InputEmail';
+import { InputMonthPage } from './generated/InputMonth';
 import { InputNumberPage } from './generated/InputNumber';
 import { InputPasswordPage } from './generated/InputPassword';
 import { InputSearchPage } from './generated/InputSearch';
@@ -119,6 +121,7 @@ import { InputTelPage } from './generated/InputTel';
 import { InputTextPage } from './generated/InputText';
 import { InputTimePage } from './generated/InputTime';
 import { InputUrlPage } from './generated/InputUrl';
+import { InputWeekPage } from './generated/InputWeek';
 import { LinkPage } from './generated/Link';
 import { LinkPurePage } from './generated/LinkPure';
 import { LinkSocialPage } from './generated/LinkSocial';
@@ -532,6 +535,11 @@ export const generatedRoutes: RouteType[] = [
     element: <Flyout_2Page />,
   },
   {
+    name: 'Flyout 3',
+    path: '/flyout-3',
+    element: <Flyout_3Page />,
+  },
+  {
     name: 'Flyout Backdrop Blur',
     path: '/flyout-backdrop-blur',
     element: <FlyoutBackdropBlurPage />,
@@ -737,6 +745,11 @@ export const generatedRoutes: RouteType[] = [
     element: <InputEmailPage />,
   },
   {
+    name: 'Input Month',
+    path: '/input-month',
+    element: <InputMonthPage />,
+  },
+  {
     name: 'Input Number',
     path: '/input-number',
     element: <InputNumberPage />,
@@ -770,6 +783,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Input Url',
     path: '/input-url',
     element: <InputUrlPage />,
+  },
+  {
+    name: 'Input Week',
+    path: '/input-week',
+    element: <InputWeekPage />,
   },
   {
     name: 'Link',