+++ packages/components/src/components/input-week/input-week-utils.ts
@@ -1,0 +1,6 @@
+import type { FormState } from '../../utils/form/form-state';
+export type InputWeekState = FormState;
+
+export type InputWeekChangeEventDetail = Event;
+export type InputWeekBlurEventDetail = Event;
+export type InputWeekInputEventDetail = InputEvent;