+++ packages/components/src/components/multi-select/multi-select/multi-select.tsx
@@ -172,11 +172,6 @@ export class MultiSelect {
 
   private currentlyHighlightedOption: Option | null = null;
 
-  @Listen('keydown')
-  public onKeyDown(e: KeyboardEvent): void {
-    this.onComboKeyDown(e);
-  }
-
   @Listen('internalOptionUpdate')
   public updateOptionHandler(e: Event & { target: MultiSelectOption }): void {
     e.stopPropagation();
@@ -261,6 +256,7 @@ export class MultiSelect {
     this.inputSearchElement = this.filterSlot
       ? (this.filterSlot.assignedElements()[0] as HTMLPInputSearchElement)
       : this.inputSearchElement;
+    this.filterSlot && this.inputSearchElement.addEventListener('keydown', this.onComboKeyDown);
     const nativeInput = this.inputSearchElement.shadowRoot.querySelector('input');
     // Avoid error in disconnectedCallback when inputSearchInputElement is not defined
     if (nativeInput) {
@@ -331,12 +327,13 @@ export class MultiSelect {
           disabled={this.disabled}
           onClick={this.onComboClick}
           onBlur={this.onComboBlur}
+          onKeyDown={this.onComboKeyDown}
           ref={(el) => (this.buttonElement = el)}
         >
           {hasCustomSelectedSlot ? (
             <slot name="selected"></slot>
           ) : (
-            <span>{this.selectedOptions.map((option) => option.textContent).join(', ')}</span>
+            <span>{this.selectedOptions.map((option) => (option.textContent ?? '').toString().trim()).join(', ')}</span>
           )}
           {this.value.length > 0 && (
             <PrefixedTagNames.pButtonPure
@@ -388,6 +385,7 @@ export class MultiSelect {
               onInput={this.onFilterInput}
               onBlur={(e: any) => e.stopPropagation()}
               onChange={(e: any) => e.stopPropagation()}
+              onKeyDown={this.onComboKeyDown}
               ref={(el: HTMLPInputSearchElement) => (this.inputSearchElement = el)}
             />
           )}